+++ chat-components/src/components/inputvalidationpane/common/default/defaultStyles/defaultInputValidationPaneCancelButtonStyles.ts
@@ -12,5 +12,6 @@ export const defaultInputValidationPaneCancelButtonStyles: IStyle = {
     fontWeight: "500",
     lineHeight: "19px",
     textOverflow: "ellipsis",
-    whiteSpace: "nowrap"
+    whiteSpace: "nowrap",
+    width: "80px"
 };